% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_requests.R
\name{add_pull_request_reviewers}
\alias{add_pull_request_reviewers}
\title{Sets review requests from a specified list of users to a pull request}
\usage{
add_pull_request_reviewers(id, users, .api_url = api_url())
}
\arguments{
\item{id}{Pull Request ID}

\item{users}{A list of users (by username) to add as reviewers from the pull request}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
The title and id of the pull request
}
\description{
Sets review requests from a specified list of users to a pull request
}
