% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_users_to_org}
\alias{add_users_to_org}
\title{Adds users to an organization}
\usage{
add_users_to_org(users, org, .api_url = api_url(graphql = FALSE))
}
\arguments{
\item{users}{A vector of usernames to add}

\item{org}{The organization to invite them to}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/" This function uses the v3 REST API instead of the v4 GraphQL}
}
\value{
A message showing the users that were added to the org
}
\description{
Adds users to an organization
}
