% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-url.R
\name{api_url}
\alias{api_url}
\title{construct an api_url}
\usage{
api_url(scheme = getOption("ghpm.scheme", "https"),
  hostname = getOption("ghpm.hostname", "api.github.com"),
  graphql = TRUE)
}
\arguments{
\item{scheme}{https or http}

\item{hostname}{hostname}

\item{graphql}{return the graphql url, false}
}
\description{
construct an api_url
}
\details{
The GraphQL endpoint
The REST API v3 has numerous endpoints; the GraphQL API v4 has a single endpoint:

http(s)://[hostname]/api/graphql
vs for v3 it is at http(s)://[hostname]/api/v3

for github.com https://api.github.com and https://api.github.com/graphql
}
