% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{assign_to_object}
\alias{assign_to_object}
\title{Assigns a list of users to an object (usually an issue or pull request)}
\usage{
assign_to_object(id, users, .api_url = api_url())
}
\arguments{
\item{id}{The id of the object}

\item{users}{A list of user IDs}
}
\description{
Assigns a list of users to an object (usually an issue or pull request)
}
