% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectboard.R
\name{create_projectboard}
\alias{create_projectboard}
\title{Creates a projectboard}
\usage{
create_projectboard(org, repo, name, body = "", columns = NULL,
  .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{name}{Name of projectboard to create}

\item{body}{Body of projectboard to create. Defaults to "".}

\item{columns}{Optional parameter to specify a vector of projectboard columns to create. (ie: `column = c('column1', 'column2', 'column3', 'column4')`). Defaults to NULL}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
Boolean value if creation was successful.
}
\description{
Creates a projectboard
}
