% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectboard.R
\name{create_projectboard_column}
\alias{create_projectboard_column}
\title{Creates a projectboard column}
\usage{
create_projectboard_column(id, name,
  .api_url = "https://api.github.com/graphql")
}
\arguments{
\item{id}{ID of the projectboard}

\item{name}{Name of the column to create}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
The ID of the projectboard the column was created in
}
\description{
Creates a projectboard column
}
