% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_requests.R
\name{create_pull_request}
\alias{create_pull_request}
\title{Creates a pull request}
\usage{
create_pull_request(org, repo, base, head, title, body = "",
  reviewers = NULL, assignees = NULL, .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{base}{Name of base branch that changes will be pulled into}

\item{head}{Name of head branch where changes will be pulled from}

\item{title}{Title of pull request}

\item{body}{Body of pull request. Defaults to ""}

\item{reviewers}{List of usernames to request reviews from (ie: `reviewers = c('devinp', 'harshb')`)}

\item{assignees}{List of usernames to assign to the pull request.}

\item{.api_url}{Optional API url to query. Defaults to the value set by `api_url()`. Usually it's "https://api.github.com/graphql"}
}
\value{
A list containing the title, creation date, author, reviewers, and assignees of the pull request
}
\description{
Creates a pull request
}
