% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{delete_all_labels}
\alias{delete_all_labels}
\title{Deletes all the labels from a given repository.}
\usage{
delete_all_labels(org, repo, .api_url = "https://api.github.com/")
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/"}
}
\value{
A message showing the tags that were deleted
}
\description{
Deletes all the labels from a given repository.
}
