% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_requests.R
\name{get_all_pull_requests}
\alias{get_all_pull_requests}
\title{Gets a data frame of the pull requests associated with a given repo}
\usage{
get_all_pull_requests(org, repo, .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the pullrequest | title | author | body | milestone | created_at | merged_by | merged_at | merged_to | state of each issue
}
\description{
Gets a data frame of the pull requests associated with a given repo
}
