% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issue_assignees}
\alias{get_issue_assignees}
\title{Gets a data frame of the assignees of each issue}
\usage{
get_issue_assignees(org, repo, .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing issue | assignedTo of each issue. Returns an empty dataframe if none are found.
}
\description{
Gets a data frame of the assignees of each issue
}
