% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_issue_events}
\alias{get_issue_events}
\title{Gets a data frame of the project board events of each issue}
\usage{
get_issue_events(org, repo, .api_url = "https://api.github.com/graphql")
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing issue | project | type | column | author | date of each issue
}
\description{
Gets a data frame of the project board events of each issue
}
