% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issues}
\alias{get_issues}
\title{Gets a data frame of the issues associated with a given repo}
\usage{
get_issues(org, repo, .api_url = "https://api.github.com/graphql")
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the number | title | body | creator | milestone | state of each issue
}
\description{
Gets a data frame of the issues associated with a given repo
}
