% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_milestones}
\alias{get_milestones}
\title{Gets a data frame of the milestones associated with a given repo.}
\usage{
get_milestones(org, repo, .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the title | description | state | author | url of each milestone
}
\description{
Gets a data frame of the milestones associated with a given repo.
}
