% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_projectboard}
\alias{get_projectboard}
\title{Gets a data frame of the issues and their columns on the project board}
\usage{
get_projectboard(org, repo, .api_url = "https://api.github.com/graphql")
}
\value{
A data frame
}
\description{
Gets a data frame of the issues and their columns on the project board
}
