% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectboard.R
\name{get_projectboard_columns}
\alias{get_projectboard_columns}
\title{Helper function for get_projectboard that returns a data frame of containing column information for each issue}
\usage{
get_projectboard_columns(.acc, .cv)
}
\arguments{
\item{.acc}{Accumlator Value}

\item{.cv}{Current Value}
}
\value{
A data frame containing issue | title | column information about an issue
}
\description{
Helper function for get_projectboard that returns a data frame of containing column information for each issue
}
