% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectboard.R
\name{get_projectboard_info}
\alias{get_projectboard_info}
\title{Gets info about a projectboard}
\usage{
get_projectboard_info(org, repo, number, .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{number}{The number of the projectboard to query}

\item{.api_url}{Optional API url to query. Defaults to the value set by `api_url()`. Usually it's "https://api.github.com/graphql"}
}
\value{
A list containing information about the projectboard
}
\description{
Gets info about a projectboard
}
