% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_projectboard_issues}
\alias{get_projectboard_issues}
\title{Helper function for get_projectboard that returns a data frame of numbers and issues}
\usage{
get_projectboard_issues(.acc, .cv)
}
\arguments{
\item{.acc}{Accumlator Value}

\item{.cv}{Current Value}
}
\value{
A data frame containing issue | title of an issue
}
\description{
Helper function for get_projectboard that returns a data frame of numbers and issues
}
