% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_requests.R
\name{get_pull_request_comments}
\alias{get_pull_request_comments}
\title{Gets a data frame of the comments of all a pull request}
\usage{
get_pull_request_comments(org, repo, number,
  .api_url = "https://api.github.com/graphql")
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{number}{The number of the pullrequest to grab data from.}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the author | body | created_at of each pull request comment. Returns an empty dataframe if none are found.
}
\description{
Gets a data frame of the comments of all a pull request
}
