% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_request_commits.R
\name{get_pull_request_commits}
\alias{get_pull_request_commits}
\title{Gets a data frame of the commits of a pull request}
\usage{
get_pull_request_commits(org, repo, number, pages = 1, .cc = FALSE,
  .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{number}{The number of the pullrequest to grab data from.}

\item{.cc}{Parse the commits as conventional commits. Defaults to FALSE.}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the oid | message | author | date of each commit of a pull request.
If .cc = TRUE, the data frame will contain oid | type | description | body | footer | author | date
}
\description{
Gets a data frame of the commits of a pull request
}
