% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_requests.R
\name{get_pull_request_reviewers}
\alias{get_pull_request_reviewers}
\title{Gets a data frame of the reviewers of all the pull requests of a given repo}
\usage{
get_pull_request_reviewers(org, repo, .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the pullrequest | reviewer. Returns an empty dataframe if none are found.
}
\description{
Gets a data frame of the reviewers of all the pull requests of a given repo
}
