% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullrequests.R
\name{get_pullrequest_comments}
\alias{get_pullrequest_comments}
\title{Gets a data frame of the comments of all the pull requests of a given repo}
\usage{
get_pullrequest_comments(org, repo,
  .api_url = "https://api.github.com/graphql")
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the pullrequest | author | body | createdAt of each pull request comment. Returns an empty dataframe if none are found.
}
\description{
Gets a data frame of the comments of all the pull requests of a given repo
}
