% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullrequest_commits.R
\name{get_pullrequest_commits}
\alias{get_pullrequest_commits}
\title{Gets a data frame of the commits of all the pull requests of a given repo}
\usage{
get_pullrequest_commits(org, repo, .cc = FALSE,
  .api_url = "https://api.github.com/graphql")
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.cc}{Parse the commits as conventional commits. Defaults to FALSE.}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the pullrequest | oid | message | author | date of each commit of a pull request.
}
\description{
Gets a data frame of the commits of all the pull requests of a given repo
}
