% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_query_results}
\alias{get_query_results}
\title{Wrapper around the get_ functions to page through results and extract the data from the response}
\usage{
get_query_results(gql_file, param_list, pages = 1, ...)
}
\arguments{
\item{gql_file}{The .graphql (with path) for the query}

\item{param_list}{List of strings, in order, specifying the property to extract from the response. For example c("repository","pullRequest","commits")}

\item{pages}{Number of pages to paginate and pull data from. Each page will contain upto 100 issues/pullrequests. Defaults to 1 page.}

\item{...}{pass through all args (named) that will be passed to graphql_query()}
}
\description{
Wrapper around the get_ functions to page through results and extract the data from the response
}
