% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghpm.R
\name{get_token}
\alias{get_token}
\title{Helper function for validating tokens}
\usage{
get_token(.api_url)
}
\arguments{
\item{.api_url}{API URL you will be querying}
}
\value{
A Personal Access Token via a system environment variable
}
\description{
Helper function for validating tokens
}
