% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_users_by_org}
\alias{get_users_by_org}
\title{Gets all the users in an organization}
\usage{
get_users_by_org(org, .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{.api_url}{Optional API url to query. Defaults to the value set by `api_url()`. Usually it's "https://api.github.com/graphql"}
}
\value{
A vector of usernames in an organization
}
\description{
Gets all the users in an organization
}
