% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghpm.R
\name{ghpm}
\alias{ghpm}
\title{GitHub Project Management}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}

\item{pages}{Number of pages to paginate and pull data from. Each page will contain upto 100 issues/pullrequests. Defaults to NULL for all pages.}
}
\description{
GHPM is a library written in R that grabs and parses data from a github repository making it easier for users to perform analytics.
Using Github's flexible GraphQL API and R's dataframe structure, ghpm returns repository data about issues, milestones, project boards, and
pull requests which can be melded together to identify useful analytical information.
}
