% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghpm.R
\name{graphql_query}
\alias{graphql_query}
\title{Parses a specified GraphQL Query from the project directory.}
\usage{
graphql_query(file, ..., .api_url = api_url(), .header = NULL)
}
\arguments{
\item{file}{Query file to execute}

\item{...}{Additional variables to pass on to the query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}

\item{.header}{Optional vector of headers to send to query}
}
\value{
The list containing the query result
}
\description{
Parses a specified GraphQL Query from the project directory.
}
