% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullrequest_commits.R
\name{parse_body}
\alias{parse_body}
\title{Parses the commit message which just contains body and footer}
\usage{
parse_body(commit_message = NULL)
}
\arguments{
\item{commit_message}{The commit message}
}
\value{
A list according to the spec body | footer
}
\description{
Parses the commit message which just contains body and footer
}
