% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_request_commits.R
\name{parse_commit}
\alias{parse_commit}
\title{Parses the commit summary and message}
\usage{
parse_commit(commit_summary = NULL, commit_message = NULL)
}
\arguments{
\item{commit_summary}{The commit summary}

\item{commit_message}{commit message}
}
\value{
A list according to spec: type | description | body | footer
}
\description{
Parses the commit summary and message
}
