% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullrequest_commits.R
\name{parse_summary}
\alias{parse_summary}
\title{Parses the commit summary which just contains type and description}
\usage{
parse_summary(commit_summary = NULL)
}
\arguments{
\item{commit_summary}{The commit summary}
}
\value{
A list according to spec: type | description
}
\description{
Parses the commit summary which just contains type and description
}
