% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullrequest_commits.R
\name{process_commit}
\alias{process_commit}
\title{Parses each commit list object and returns a single row data.frame}
\usage{
process_commit(.x)
}
\arguments{
\item{.x}{a commit object}
}
\value{
A data.frame with the following structure type | description | body | footer
}
\description{
Parses each commit list object and returns a single row data.frame
}
