% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_response}
\alias{sanitize_response}
\title{Sanitizes the response from the `graphql_query()` function.}
\usage{
sanitize_response(response, stop = TRUE)
}
\arguments{
\item{response}{The result of `graphql_query()`}

\item{stop}{If true, will stop the execution when an error is returned otherwise returns the response.}
}
\description{
Sanitizes the response from the `graphql_query()` function.
}
