% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectboard.R
\name{create_projectboard}
\alias{create_projectboard}
\title{Creates a projectboard}
\usage{
create_projectboard(org, repo, name, body = "",
  .api_url = "https://api.github.com/graphql")
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{name}{Name of projectboard to create}

\item{body}{Body of projectboard to create}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
The ID of the projectboard that was created
}
\description{
Creates a projectboard
}
