% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issue_events}
\alias{get_issue_events}
\title{Gets a data frame of the project board events of each issue}
\usage{
get_issue_events(org, repo, .api_url = api_url(), pages = 1)
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}

\item{pages}{Number of pages to paginate and pull data from. Each page will contain upto 100 issues/pullrequests. Defaults to 1 page.}
}
\value{
A data frame containing issue | project | type | column | author | date of each issue. Returns an empty dataframe if none are found.
}
\description{
Gets a data frame of the project board events of each issue
}
