% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issues}
\alias{get_issues}
\title{Gets a data frame of the issues associated with a given repo}
\usage{
get_issues(org, repo, .api_url = api_url(), pages = NULL)
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to the value set by `api_url()`. Usually it's "https://api.github.com/graphql"}

\item{pages}{Number of pages to paginate and pull data from. Each page will contain upto 100 issues/pullrequests. Defaults to NULL for all pages.}
}
\value{
A data frame containing the issue | title | body | creator | milestone | state of each issue
}
\description{
Gets a data frame of the issues associated with a given repo
}
