% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectboard.R
\name{get_projectboard}
\alias{get_projectboard}
\title{Gets a data frame of the issues and their columns on the project board}
\usage{
get_projectboard(org, repo, .api_url = "https://api.github.com/graphql")
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
A data frame containing the issue | title | column | board of the project boards
}
\description{
Gets a data frame of the issues and their columns on the project board
}
