% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectboard.R
\name{get_projectboard_issues}
\alias{get_projectboard_issues}
\title{Gets a data frame of the issues and their columns on the project board}
\usage{
get_projectboard_issues(org, repo, .api_url = api_url())
}
\arguments{
\item{org}{Name of organization to query}

\item{repo}{Name of repository to query}

\item{.api_url}{Optional API url to query. Defaults to the value set by `api_url()`. Usually it's "https://api.github.com/graphql"}
}
\value{
A data frame containing the board | column | title | issue of the projectboards
}
\description{
Gets a data frame of the issues and their columns on the project board
}
