% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_user_info}
\alias{get_user_info}
\title{Gets data about a user based on their login name.}
\usage{
get_user_info(username, .api_url = api_url())
}
\arguments{
\item{username}{The user's login name}

\item{.api_url}{Optional API url to query. Defaults to the value set by `api_url()`. Usually it's "https://api.github.com/graphql"}
}
\value{
A list containing data about the user
}
\description{
Gets data about a user based on their login name.
}
