% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{graphql_query}
\alias{graphql_query}
\title{Parses a specified GraphQL Query from the project directory.}
\usage{
graphql_query(file, ...,
  .api_endpoint = "https://api.github.com/graphql")
}
\arguments{
\item{file}{Query file to execute}

\item{...}{Additional variables to pass on to the query}

\item{.api_url}{Optional API url to query. Defaults to "https://api.github.com/graphql"}
}
\value{
The list containing the query result
}
\description{
Parses a specified GraphQL Query from the project directory.
}
